Removed all back layers from glyphs that had them.
In most cases, this layer had been used as a guide, slightly modified to
make the front layer.  Sometimes it was just a mistake.

This remedies
bug #22381: Some characters have "ghost" glyphs in back layer
https://savannah.gnu.org/bugs/index.php?22381

However, this is bound to happen again.  It would be good to have a script
that checks, or maybe a facility in FontForge's "Find Problems".