Clean-up of problems found by FontForge "Find Problems".
Fixed all problems with
	Open Paths
	Intersecting Paths
	Check outermost paths clockwise
With
	Check missing extrema
there were lots of problems, and it's very fiddly.  I got up through
the Arabic range, but pooped out in the Indian scripts.

Note: Most of the problems were plainly mistakes, or problems that cropped
up after some transformation of the glyph.  Others seem to be the result
of freehand drawing or scanning.

More info at:
bug #22454: Path bugs in FreeSerif