Supplemental Math operators ============== Added a few of the "n-ary" operators.
Math operators
==============
Corrected n-ary union, intersection, and spikes to be larger than the
binary operators

Made (many of) the operators based on + - = to use those symbols directly
(by reference or copying).
I think it is good practice to harmonize the appearance of math
operators with the Latin character set.

Geometrical Shapes
==================
Made the large circle larger still.  When is it large enough?

Combining marks for Symbols
===========================

Made the combining circle to really combine.
Made a circle-slash