Think I fixed the problem causing the FontForge crashes.  Related to the dotlessi thing.
There's a table
Lookup: 6 0 0 "'ccmp' Glyph Composition/Decomposition in Latin lookup 0"  {"'ccmp' Glyph Composition/Decomposition in Latin lookup 0 subtable"  } ['ccmp' ('latn' <'dflt' > ) ]
It is related to another lookup
ChainSub2: class "'ccmp' Glyph Composition/Decomposition in Latin lookup 0 subtable"
that has something to do with i and j.  It ends with
  SeqLookup: 0 "Single Substitution lookup 47"
  EndFPST
When I removed the lookup for i and j, the SeqLookup line would be badly
corrupted, and subsequent saves resulted in a segfault and a truncated
sfd file.

So I doctored the whole thing with vi.