Added at 0x0000 0x0001 0x000D characters          .notdef .null nonmarkingreturn These are NOT UNICODE they are to satisfy an essentially undocumented "feature" of FontForge, ( in source tottf.c, AssignNotdefNull() ) And a funny strong suggestion of TrueType fonts, that they begin with three special characters (evidently different from encoding slots) http://www.microsoft.com/typography/otspec/recom.htm
FontForge looks for these glyphs in certain places and copies them.
Otherwise, apparently, it will find a glyph elsewhere (unclear how) and
copy it there, possibly resulting in something quite wrong
(such as a non-monospace font).

The three are copied to the end of TTF fonts (only one to the end of OTF
fonts)