Glyph cleanup: Addressing the usual pointwise problems, but also the 	TrueType restriction that references be offset by integral values.
The bulk of the remaining test warnings concern (ATT) complaints about
tables with characters in Indic ranges not being in the proper script for 
the range.  This is evidently a FontForge bug:  some character ranges have
multiple scripts -- several Indic ranges do.  A misunderstanding like that
would explain it.  Another group is about the use of 'danda' in 'locl', 
again I think it's a confusion about the relation of scripts and ranges.
One more problem with the new range 1DC0 -- the complaint is that tables
with those characters are being applied with non-latin scripts: but this
range is not associated with latin.

Also there are several complaints about glyphs being named incorrectly.
In each case FontForge is simply wrong.  Just a bug.

The infamous "invalid PS name" complaint remains: it's a FontForge bug, but 
what on earth is triggering it?  Plan is to get rid of that.

One other issue: in making the serif oblique Mathematical Operators ranges,
I used references, forgetting this would cause a missing extrema complaint.