Latin, Cyrillic ranges: letters with bar / horizontal slash.  Toward 	bug #50949: Stroke overlays in small caps.
The problem that is more than aesthetic was the distinction between
ұ (uk_precherknutyj): both this and ɏ (ystroke) get mapped to small cap y,
although the bar of the former should be on the stem, while on the latter
it should be across the arms.  Solution: separate small caps versions.

Also normalized thicknesses of bars/strokes in both small and caps versions
of all the letters.